contop-server/pyproject.toml
[project]
name = "contop-server"
version = "0.1.0"
description = "Contop server - cross-platform desktop automation via WebRTC"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiortc>=1.14.0",
    "cryptography>=46.0.5",
    "fastapi>=0.135.1",
    "google-adk>=1.25.0",

    "mss>=10.1.0",
    "openpyxl>=3.1.0",
    "pillow>=12.1.1",
    "pyautogui>=0.9.54",
    "pymupdf4llm>=0.0.17",
    "pyperclip>=1.8.0",
    "qrcode>=8.2",
    "uvicorn>=0.41.0",
    "websockets>=15.0.1",
    "docker>=7.1.0",
    "litellm>=1.50.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
windows = [
    "pywinauto>=0.6.9",
]
macos = [
    "pyobjc-framework-Accessibility>=10.0",
]
# linux: pyatspi is a system package (apt install python3-pyatspi), not available on PyPI
omniparser = [
    "torch==2.11.0",
    "torchvision==0.26.0",
    "transformers>=4.40.0,<5.0.0",
    "ultralytics>=8.3.70",
    "easyocr>=1.7.0",
    "huggingface-hub>=0.20.0",
    "accelerate>=0.25.0",
    "timm>=0.9.0",
    "einops>=0.7.0",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.8.0",
]
cu126 = [  # triggers uv.sources to use CUDA index for torch
    "torch==2.11.0",
    "torchvision==0.26.0",
]
cpu = [    # triggers uv.sources to use CPU index for torch
    "torch==2.11.0",
    "torchvision==0.26.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.5",
]

[tool.setuptools.packages.find]
include = ["core*", "tools*", "platform_adapters*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
markers = [
    "unit: Unit tests (fast, no external deps)",
    "integration: Integration tests (may need running services)",
    "api: API endpoint tests",
]
asyncio_mode = "auto"

[tool.uv]
# cu126 and cpu extras are mutually exclusive - you install one or the other
conflicts = [[{ extra = "cu126" }, { extra = "cpu" }]]

# Use PyTorch CUDA index for torch/torchvision so GPU wheels are resolved
# automatically. Other packages still come from PyPI.
[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu126", extra = "cu126" },
    { index = "pytorch-cpu", extra = "cpu" },
]
torchvision = [
    { index = "pytorch-cu126", extra = "cu126" },
    { index = "pytorch-cpu", extra = "cpu" },
]
